+++ task11_test.go
@@ -1,0 +1,14 @@
+package main
+
+import "testing"
+
+func TestCheckPalindrom(t *testing.T) {
+	s:="aabaa"
+	expected:=true
+
+	result:=CheckPalindrom(s)
+
+	if expected!=result {
+		t.Errorf("incorrect. Expected %v, got %v",expected,result)
+	}
+}
